# openApi/apiSpec.v1.yaml
openapi: 3.0.0
info:
  contact:
    name: karmy
  title: apiSpec
  version: '1.0'
  description: Сервер авторизации
servers:
  - url: 'http://localhost:3000'
paths:
...
  /brands:
    get:
      summary: Получить данные из таблицы brand
      tags:
        - device
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    brand:
                      type: string
      operationId: get-brands
      description: 'Получает все данные из таблицы, если задан catId ответ будет содержать только колонки связанные с отправленной категорией'
      parameters:
        - schema:
            type: integer
          in: query
          name: catId
          description: ID категори
    post:
      summary: Добавить данные в таблицу brand
      operationId: post-brands
      responses:
        '200':
          description: Данные добавленной стоки
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  brand:
                    type: string
                required:
                  - id
                  - brand
      tags:
        - device
      description: Добавляет новую строку в таблицу
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                brand:
                  type: string
              required:
                - brand
        description: ''
    patch:
      summary: Редактировать данные в таблице brand
      operationId: patch-brands
      responses:
        '200':
          description: Данные измененной стоки
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  brand:
                    type: string
      tags:
        - device
      description: Редактирует запись
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                brand:
                  type: string
              required:
                - id
                - brand
        description: ''
    parameters: []
    delete:
      summary: Удлаить строку с таблицы brand
      operationId: delete-brands
      responses:
        '200':
          description: id уделенной строки
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: id удаленного устройства
                required:
                  - id
      description: В бд определенны внешние ключи которые не дадут удалить запись если она связанна с каким нибудь оборудованием
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
        description: ''
      tags:
        - device
  /suppliers:
    get:
      summary: Получить данные из таблицы supplier
      tags:
        - device
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    supplier:
                      type: string
      operationId: get-suppliers
      description: 'Получает все данные из таблицы, если задан catId ответ будет содержать только колонки связанные с отправленной категорией'
      parameters:
        - schema:
            type: integer
          in: query
          name: catId
          description: ID категори
    post:
      summary: Добавить данные в таблицу supplier
      operationId: post-suppliers
      responses:
        '200':
          description: Данные добавленной стоки
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  supplier:
                    type: string
                required:
                  - id
                  - supplier
      tags:
        - device
      description: Добавляет новую строку в таблицу
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                supplier:
                  type: string
              required:
                - supplier
        description: ''
    patch:
      summary: Редактировать данные в таблице supplier
      operationId: patch-suppliers
      responses:
        '200':
          description: Данные измененной стоки
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  supplier:
                    type: string
      tags:
        - device
      description: Редактирует запись
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                supplier:
                  type: string
              required:
                - id
                - supplier
        description: ''
    parameters: []
    delete:
      summary: Удлаить строку с таблицы supplier
      operationId: delete-suppliers
      responses:
        '200':
          description: id уделенной строки
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: id удаленного устройства
                required:
                  - id
      description: В бд определенны внешние ключи которые не дадут удалить запись если она связанна с каким нибудь оборудованием
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      tags:
        - device
...
  /account_owners:
    get:
      summary: Получить список владельцев сериса
      tags:
        - accounts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    user_id:
                      type: integer
                    dep_loc_id:
                      type: integer
                    account_id:
                      type: integer
                  required:
                    - id
                    - user_id
                    - dep_loc_id
                    - account_id
      operationId: get-account_owners
      description: Получить список владельцев сериса
      parameters:
        - schema:
            type: integer
          in: query
          name: accountId
        - schema:
            type: integer
          in: query
          name: depLocId
        - schema:
            type: integer
          in: query
          name: userId
    post:
      summary: Добавить владельца сервису
      operationId: post-account_owners
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  dep_loc_id:
                    type: integer
                  account_id:
                    type: integer
                required:
                  - id
                  - user_id
                  - dep_loc_id
                  - account_id
      tags:
        - accounts
      description: Добавить владельца сервису
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                dep_loc_id:
                  type: integer
                account_id:
                  type: integer
              required:
                - account_id
    delete:
      summary: Удалить владельца у сервиса
      operationId: delete-account_owners
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    user_id:
                      type: integer
                    dep_loc_id:
                      type: integer
                    account_id:
                      type: integer
                  required:
                    - id
                    - user_id
                    - dep_loc_id
                    - account_id
      description: Удалить владельца у сервиса
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                account_id:
                  type: integer
                user_id:
                  type: integer
                dep_loc_id:
                  type: integer
      tags:
        - accounts
...
  /mentoring:
    get:
      summary: Просмотр наставничества
      tags:
        - HR
      responses:
        '200':
          description: OK
      operationId: get-mentoring
      description: Просмотр наставничества
    post:
      summary: Создание наставничества
      operationId: post-mentoring
      responses:
        '200':
          description: OK
      description: Создание наставничества
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mentor_id:
                  type: integer
                protege_id:
                  type: integer
              required:
                - mentor_id
                - protege_id
      tags:
        - HR
    patch:
      summary: Создание плана
      operationId: patch-mentoring
      responses:
        '200':
          description: OK
      description: Создание плана
      tags:
        - HR
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                plan:
                  type: object
                  nullable: true
                  properties:
                    blocks:
                      type: array
                      items:
                        type: object
                        properties:
                          title:
                            type: string
                          sections:
                            type: array
                            items:
                              type: object
                              properties:
                                title:
                                  type: string
                          test:
                            type: object
                            properties:
                              title:
                                type: string
                              questions:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    title:
                                      type: string
                                    answers:
                                      type: array
                                      items:
                                        type: object
                                        properties:
                                          title:
                                            type: string
                                          img:
                                            type: string
                                          isRight:
                                            type: integer
                                            enum:
                                              - 0
                                              - 1
                                          protegeСhoice:
                                            type: integer
                                    img:
                                      type: string
                          task:
                            type: object
                            properties:
                              title:
                                type: string
    put:
      summary: Отправить на проверку всеми
      operationId: put-mentoring
      responses:
        '200':
          description: OK
      description: Отправить на проверку всеми
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
      tags:
        - HR
...
components:
...
  securitySchemes: {}
...